+++ src/cli/command-parser.ts
@@ -3,19 +3,17 @@ type ParsedCommand = Record<string, string[]>
 export class CommandParser {
   static parse(cliArguments: string[]): ParsedCommand {
     const parsedCommand: ParsedCommand = {};
-    let currentCommand: any;
+    let currentCommand = '';
 
     for (const argument of cliArguments) {
       if (argument.startsWith('--')) {
-
         parsedCommand[argument] = [];
         currentCommand = argument;
-
       } else if (currentCommand && argument) {
         parsedCommand[currentCommand].push(argument);
       }
     }
-    
+
     return parsedCommand;
   }
 }
