+++ src/cli/commands/command.interface.ts
