+++ src/main.cli.ts
@@ -8,7 +8,7 @@ function bootstrap() {
     new VersionCommand(),
     new ImportCommand(),
   ]);
-  
+
   cliApplication.processCommand(process.argv);
 }
