+++ src/shared/types/city.type.ts
@@ -1,0 +1,1 @@
+export type City = 'Paris' | 'Cologne ' | 'Brussels ' | 'Amsterdam ' | 'Hamburg' | 'Dusseldorf ';
