+++ src/shared/types/comment.type.ts
@@ -1,0 +1,8 @@
+import { User } from './index.js';
+
+export type Comment = {
+    content: string,
+    postDate: Date,
+    rating: number,
+    user: User
+}
