+++ src/shared/types/good.type.ts
@@ -1,0 +1,1 @@
+export type Good = 'Breakfast' | 'Air conditioning' | 'Laptop friendly workspace' | 'Baby seat' | 'Washer' | 'Towels' | 'Fridge';
