+++ src/shared/types/index.ts
@@ -1,1 +1,7 @@
-export * from './offers.type.js';
+export * from './offer.type.js';
+export * from './offer-type.type.js';
+export * from './location.type.js';
+export * from './city.type.js';
+export * from './good.type.js';
+export * from './user.type.js';
+export * from './comment.type.js';
