+++ src/shared/types/location.type.ts
@@ -1,0 +1,4 @@
+export type Location = {
+    latitude: number
+    longitude: number
+}
