+++ src/shared/types/offer-type.type.ts
@@ -1,0 +1,1 @@
+export type OfferType = 'apartment' | 'house' | 'room' | 'hotel';
