+++ src/shared/types/user.type.ts
@@ -1,0 +1,7 @@
+export type User = {
+    avatarUrl: string,
+    isPro: boolean,
+    name: string,
+    mail: string,
+    password: string
+}
